pyproject.toml
[tool.poetry]
name = "perm"
version = "0.1.0"
description = ""
authors = ["hggforget <1242537651@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
antlr4-python3-runtime = "4.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
